requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.7
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sympy==1.12
threadpoolctl==3.4.0
timm==0.9.16
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.2
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
yacs==0.1.8
